// web/lib/api/generated/model/adjustmentLine.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */
import type { AdjustmentModeEnum } from './adjustmentModeEnum';
import type { StockProductRef } from './stockProductRef';

export interface AdjustmentLine {
  readonly id: string;
  readonly line_no: number;
  readonly product: StockProductRef;
  readonly mode: AdjustmentModeEnum;
  /** @pattern ^-?\d{0,11}(?:\.\d{0,3})?$ */
  readonly entered_qty: string;
  /** @pattern ^-?\d{0,11}(?:\.\d{0,3})?$ */
  readonly quantity_change: string;
  /** @pattern ^-?\d{0,11}(?:\.\d{0,3})?$ */
  readonly on_hand_before: string;
}