// web/lib/api/generated/model/adjustmentModeEnum.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `ADD` - Add
 * * `REMOVE` - Remove
 * * `COUNTED` - Counted
 */
export type AdjustmentModeEnum = typeof AdjustmentModeEnum[keyof typeof AdjustmentModeEnum];


export const AdjustmentModeEnum = {
  ADD: 'ADD',
  REMOVE: 'REMOVE',
  COUNTED: 'COUNTED',
} as const;