// web/lib/api/generated/model/reasonCodeEnum.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `COUNT_CORRECTION` - Stock count correction
 * * `DAMAGE` - Damaged
 * * `EXPIRY` - Expired
 * * `THEFT` - Lost or stolen
 * * `OPENING_STOCK` - Opening stock
 * * `OTHER` - Other
 */
export type ReasonCodeEnum = typeof ReasonCodeEnum[keyof typeof ReasonCodeEnum];


export const ReasonCodeEnum = {
  COUNT_CORRECTION: 'COUNT_CORRECTION',
  DAMAGE: 'DAMAGE',
  EXPIRY: 'EXPIRY',
  THEFT: 'THEFT',
  OPENING_STOCK: 'OPENING_STOCK',
  OTHER: 'OTHER',
} as const;